+++ src/connection.ts
@@ -1,12 +1,17 @@
 import knex from "knex";
+import dotenv from "dotenv";
 
-export const connection = knex({
-    client: "mysql",
-    connection: {
-      host: process.env.DB_HOST,
-      port: 3306,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_NAME,
-    },
-  });
+dotenv.config();
+
+const connection = knex({
+  client: "mysql",
+  connection: {
+     host: process.env.DB_HOST,
+     port: 3306,
+     user: process.env.DB_USER,
+     password: process.env.DB_PASSWORD,
+     database: process.env.DB_NAME
+  }
+})
+
+  export default connection
