+++ src/endpoints/createStudent.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { connection } from "../connection";
+import connection from "../connection";
 
 export const createStudent = async (req: Request, res: Response): Promise<void> => {
